semi commit

+++ StepCounterApp/src/main/java/com/greenland/stepcounter/stepvalue/Values.java
@@ -3,5 +3,5 @@ package com.greenland.stepcounter.stepvalue;
 public class Values {
 	public static int Step = 0;
 	public static String Address = new String();
-	public static float Distance = 0;
+	public static int Distance = 0;
 }
